refactor(server): 폴더명 변경
