+++ server/src/core/verifyUser.ts
